requirements.txt
transformers==4.44.2
accelerate>=0.33.0
sentencepiece>=0.2.0
ultralytics==8.3.20
pillow>=10.4.0
numpy>=1.26.0
opencv-python>=4.10.0.84
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
python-multipart>=0.0.9
jupyterlab>=4.2.5
ipykernel>=6.29.5

# IMPORTANT: Install PyTorch separately per OS/compute from https://pytorch.org/get-started/locally/
# Example CPU on Windows:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
